fix test case of file `helperFunctions.test.ts`
Two test cases were failing because of change in `month` and `year` value
respectively. In function `extractCounter` in case of monthly we're checking
current month and in case of year we're checking current year. As part of fix
in the test case I'm considering current month and year so that it should pass
the condition.